Move tranpose test to higher level pairing tests.
We have to change the pairing internals. We can't use the transpose
method, because the number of times we call it is O(n!), and no matter
how fast we make that method, sooner of later we're going to call it
thousands of millions of times.

So, since we're going to break the tests regarding the tranpose methods,
we use the same logic to test the method returning the pairs.