Remove dependency between bracket and player.
The bracket actually doesn't require Player objects, just objects
responding to "number", "points", "<=>" and "compatible_with?".

However, the tests are easier to write if we create real Player objects,
because the "compatible_with?" method is very difficult to stub with
doubles and the pairing mechanishm is highly dependent on this method.