Remove duplication initializing methods.
I'm not sure this is a good move, because we're adding behavior to the
Class class and the duplication we remove with the method, we add it by
requiring the same file from every class.

So hopefully there are better ways, but for now, let's experiment. I like
the feeling of writing things like initialize_with :players.