Require bracket instead of exchanged_bracket.
The ExchangedBracket class is a class whose existence shouldn't be known
by any other classes other than the Bracket class.

For any other classes, the Bracket class should be the class we require.

The reason for that is to keep the interface clean and consistent with
the real world. A Bracket is something that exists in the real world,
but an ExchangedBracket is something I made up which was helpful to
organize code in the Bracket class.