Remove ExchangedBracket class.
It simplifies the code in some ways, particularly the bits related to s1
and s1, and it makes it also clear that we can exchange while we don't
find pairs.

It also avoids calculations of every possible exchange and its pairs,
although I haven't detected any performance improvement

The main inconvinient is we now need to track how many exchanges we've
done.