Add class to handle exchanges.
Now we remove more logic from the Bracket class, and we don't have to
rely on extending a core class like we did with ArrayExchanger.

The tests now seem to be the same as the tests for Bracket#exchange.
However, the tests for Bracket#exchange also check that S1 and S2 are
correctly reordered, so I'm leaving just one of those tests.