Fix exchanges in heterogeneous brackets.
We weren't removing elements from S1 when adding elements to Limbo,
resulting in unexpected results. So instead of treating Limbo as a
separate entity, we use the same logic as we did for homogeneous
bracket, only a heterogeneous bracket consists of S1+Limbo+S2, with
exchanges taking place between S1 and Limbo.

These exchanges follow the rules defined in section D.3.