Change homogeneous bracket definition.
It was updated in 2016, section A.3: "A (pairing) bracket is homogeneous
if all the players have the same score; otherwise it is heterogeneous."

It affects some cases where according to the old definition a bracket
would be homogeneous, but they would be heterogeneous after some moved
down players are marked to downfloat again. I've added a test case.

It also affects the Collapsed Last Bracket, but that one isn't
implemented yet. For now, we're still considering it homogenous.