Don't sort S2 players.
I thought some tests might break, but the algorithm isn't affected
because once a exchange is done, we've already checked the cases where
S2 is sorted.

And the number of calls to Array#sort is one of our bottlenecks, so now
tests run 10% faster, and performance tests pass on my machine again.