Test special CLB pairings.
When we find the last two brackets can't be paired, we downfloat players
from the previous bracket which allow pairing completion. So, in the
example, we downfloat 1 and 2.

If we followed section A.9, we would now form a bracket with all the
unpaired players.

Then we would try to apply criteria C.6 and pair players with the same
score against each other. This would be a special case where we try to
pair moved down players against each other. I can't find any mention to
this possibility anywhere in the FIDE Swiss Rules.

So for now I'm ignoring the CLB conditions, and just downfloating the
moved down players again, which produces the same results.