Make defective code more obvious.
So the code wasn't doing what I thought it was doing due to a bug while
checking for impossible downfloats. Basically, when we add impossible
downfloats individually, we only check correctly for them in the method
Bracket#exchange_until_s2_players_can_downfloat.

It's working because the only affected place is the penultimate bracket;
in the rest of the brackets, we don't use the method
number_of_required_downfloats.