Differenciate forbidden and impossible downfloats.
Impossible downfloats are collections of players. Forbidden downfloats
are individual players from the PPB who can't be downfloated because the
next bracket (the last one) would be impossible to pair.

There was a bug in the code making forbidden downfloats only work for
the penultimate bracket, but it turned out to be a reasonable approach;
while ideally we would handle them together, one of our existing test
cases showed a confusing situation if we did so.

Players 1 to 3 have two points, 3 to 7 have one point, and 8 to 10 have
zero points. Player 1 has played against players 2, 3, 4, 5, 6, and 7.

If we treated impossible downfloats as forbidden downfloats, we couldn't
downfloat player 1 because it has no possible opponent in the next
bracket.