Remove redundant impossible to pair check.
We're already doing that check in the previous scoregroup, so the
generated scoregroup will always be possible to pair.

The only exception would take place when the whole round is impossible
to pair; we're not supporting that case for now, but if we do, we'd add
code to the Round class.