Maximize downfloat quality criterias (C.12-C.15).
We were doing an all or nothing quality check: if we couldn't, for
example, avoid having a player downfloating twice in a row, we didn't
care for how many players downfloated twice.

But, actually, we need to try to downfloat only one player twice, and if
we fail, only two, and so on.

As expected, performance decreases because of more quality checks.