Rename OpponentsIncompatibilities to Completion.
"There are only two hard things in Computer Science: cache invalidation
and naming things." -- Phil Karlton

This class is particularly hard to name. Maybe it's because it now does
two things: it checks whether a group of players can complete the
pairing, and how many players are impossible to pair. We might split it
in the future.