Differenciate between hypothetical and provisional.
So, "provisional leftovers" refer to the leftovers in a bracket after
completing the pairing in that bracket but before checking its quality,
as opposed to "definitive pairs".

On the other hand, "hypothetical leftovers" refer to the leftovers in a
scoregroup if we take the definitive pairs the bracket has generated.