Store failing criteria while establishing pairs.
In a certain we didn't know why the pairing had failed; it happened when
established pairs were empty. It could happen because all pairs were not
ideal, or because they failed the preventive quality check.

So now, every time a quality check fails, we store why it happened.
That way, we can know which criterion we should relax in order to
complete the pairing.