Change the way we calculate floats.
There are two reasons for this change.

The first one is performance; the test tournament is now 25% faster.

The second one is about design; when we initialize a Pair, we've got
access to the player, the opponent, and the points they've got both
before and after the match. However, when we initialize a Game object,
we need to find another Game object belonging to the current game's
opponent, and compare the points they had before the match started. It
seems to me we were making things harder than they were.

On the other hand, the new code doesn't look great either.