Cache Game#played? and Game#opponent.
For now, we're improving performance the easy way. The better way would
be to improve the algorithm so these methods aren't called millions of
times. We might do it if necessary once we remove the noise these
methods make when using a profiler.

The test tournament is now 25% faster. By caching several methods, we've
reduced the runtime from about 45-50 seconds to 0.85 seconds.