Improve possible pairs performance.
Every time we find an incompatibility, we check again whether the rest
of the players can be paired without incompatibilities.

Some of our slowest tests run faster, making a 20% overal improvement
for the whole test suite. The added performance test previously took
forever to execute and now it runs instantly.