Cache player compatibilities.
The slowest test in our test suite is now 30% faster, two of our three
slowest performance tests run in half the time, and our tournament spec
is 10% faster.

I'm not sure these small improvements justify this change, but I'm
keeping it because they can help finding a bigger performance
improvement in the future.