Split QualityCriteria in two classes.
One class checks the quality of certain pairs, and one class tracks the
failing criterion and handles the allowed failures.

This way we remove duplication between the methods to check the quality
of the established pairs and the methods to check the quality of the
provisional pairs.

The main disadvantage is we need to track que failing criteria in many
classes again.