Don't run performance tests in the rake spec task.
These tests depend too much on the environment they're running on, and
are there just to manually check performance before a release. They're
not supposed to be run during normal development, since they're slow.

I had already excluded them in my .rspec file, but since that file isn't
in version control, it didn't work for other people.

Thanks to David for the tip.